requirements.txt
Flask
pdfplumber
PyPDF2
python-docx
scikit-learn

